+++ src/app/app-routing.module.ts
@@ -7,6 +7,7 @@ import { ClientsComponent } from './pages/clients/clients.component';
 import { ProfileComponent } from './pages/profile/profile.component';
 import { IsLoggedGuard } from './guards/is-logged.guard';
 import { EditprofileComponent } from './pages/profile/editprofile/editprofile.component';
+import { AdminMenuComponent } from './pages/admin-menu/admin-menu.component';
 
 const routes: Routes = [
   { path: '', component: DashboardComponent, data: {title: 'Dashboard'}},
@@ -15,6 +16,7 @@ const routes: Routes = [
   { path: 'tasks', component: TasksComponent, data: {title: 'Tasks'}, canActivate: [IsLoggedGuard]},
   { path: 'profile', component: ProfileComponent, data: {title: 'Profile'}, canActivate: [IsLoggedGuard]},
   { path: 'edit-profile', component: EditprofileComponent, data: {title: 'Edit Profile'}, canActivate: [IsLoggedGuard]},
+  { path: 'admin-menu', component: AdminMenuComponent, data: {title: 'Admin Menu'}, canActivate: [IsLoggedGuard]},
 
 
 ];